pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chembl-pdb-linker"
version = "0.1.0"
description = "Link ChEMBL bioactivity data with PDB structural information"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
]
keywords = ["chembl", "pdb", "bioactivity", "drug-discovery", "cheminformatics"]

dependencies = [
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "tqdm>=4.66.0",
    "rdkit>=2023.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
chembl-pdb-linker = "chembl_pdb_linker.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/chembl-pdb-linker"
Repository = "https://github.com/yourusername/chembl-pdb-linker"

[tool.hatch.build.targets.wheel]
packages = ["src/chembl_pdb_linker"]

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".eggs",
    "*.egg",
    "dist",
    "build",
    "data",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["B008"]  # typer.Option() in defaults is required by typer

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
pythonVersion = "3.9"
typeCheckingMode = "basic"
include = ["src"]
exclude = [
    "**/__pycache__",
    "**/.*",
    "tests",
    "data",
    "scripts",
]
reportMissingImports = false  # Disabled for pre-commit compatibility
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
reportDuplicateImport = true
reportOptionalSubscript = true
reportOptionalMemberAccess = true
reportOptionalCall = true
reportOptionalIterable = true
reportOptionalContextManager = true
reportOptionalOperand = true
reportGeneralTypeIssues = true
reportPrivateImportUsage = true
reportUnknownMemberType = false  # Disabled due to pandas type stubs limitations
reportUnknownArgumentType = false  # Disabled due to pandas type stubs limitations
reportUnknownVariableType = false  # Disabled due to pandas type stubs limitations
reportUnknownLambdaType = false
reportUnknownParameterType = false
reportMissingParameterType = false
reportUntypedBaseClass = false
reportUntypedClassDecorator = false
reportUntypedFunctionDecorator = false
reportUntypedNamedTuple = false
reportCallInDefaultInitializer = false  # typer.Option() requires this pattern
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportAssertAlwaysTrue = true
reportSelfClsParameterName = true
reportImplicitStringConcatenation = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/chembl_pdb_linker"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
